3. store_to_database/requirements.txt
python-dotenv==1.0.1
pandas==2.2.2
mysql-connector-python==9.0.0
tqdm==4.66.4
